SYAL-554 fix: add method functionality to java docs, change dependency scope of httpclient, update version of jackson-core dependency
