+++ pom.xml
@@ -63,7 +63,7 @@
         <version>4.9.10</version>
         <executions>
           <execution>
-            <id>get-the-git-infos</id>
+            <id>get-the-git-info</id>
             <goals>
               <goal>revision</goal>
             </goals>
@@ -123,13 +123,13 @@
       <groupId>com.fasterxml.jackson.core</groupId>
       <artifactId>jackson-core</artifactId>
       <scope>provided</scope>
-      <version>2.11.1</version>
+      <version>2.12.4</version>
     </dependency>
     <dependency>
       <groupId>com.fasterxml.jackson.core</groupId>
       <artifactId>jackson-databind</artifactId>
       <scope>provided</scope>
-      <version>2.11.1</version>
+      <version>2.12.4</version>
     </dependency>
     <dependency>
       <groupId>org.springframework</groupId>
@@ -142,6 +142,12 @@
       <artifactId>jaxb-api</artifactId>
       <scope>provided</scope>
       <version>2.3.1</version>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.httpcomponents</groupId>
+      <artifactId>httpclient</artifactId>
+      <scope>provided</scope>
+      <version>4.5.13</version>
     </dependency>
     <!--scope test-->
     <dependency>
@@ -192,23 +198,18 @@
       <scope>test</scope>
       <version>0.14.0</version>
     </dependency>
-    <!--scope compile-->
-    <dependency>
-      <groupId>com.google.code.gson</groupId>
-      <artifactId>gson</artifactId>
-      <version>2.8.5</version>
-      <scope>compile</scope>
-    </dependency>
     <dependency>
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
       <version>4.13.1</version>
       <scope>test</scope>
     </dependency>
+    <!--scope compile-->
     <dependency>
-      <groupId>org.apache.httpcomponents</groupId>
-      <artifactId>httpclient</artifactId>
-      <version>4.5.4</version>
+      <groupId>com.google.code.gson</groupId>
+      <artifactId>gson</artifactId>
+      <version>2.8.5</version>
+      <scope>compile</scope>
     </dependency>
   </dependencies>
 </project>
